Move signals to model file per Django docs recommendation.
